fix(windows): try to hide the entire Windows client from Ubuntu CI. It should get static analysis on a Windows runner anyway.
