+++ rust/windows-client/src-tauri/src/client/cli.rs
